+++ lib/rubocop/cop/layout/empty_lines_around_access_modifier.rb
@@ -29,6 +29,26 @@ module RuboCop
         MSG_BEFORE_AND_AFTER = 'Keep a blank line before and after ' \
                                '`%<modifier>s`.'.freeze
 
+        def on_class(node)
+          _name, superclass, _body = *node
+
+          @class_or_module_def_line = if superclass
+                                        superclass.first_line
+                                      else
+                                        node.source_range.first_line
+                                      end
+        end
+
+        def on_module(node)
+          @class_or_module_def_line = node.source_range.first_line
+        end
+
+        def on_sclass(node)
+          self_node, _body = *node
+
+          @class_or_module_def_line = self_node.source_range.first_line
+        end
+
         def on_send(node)
           return unless node.bare_access_modifier?
 
@@ -40,11 +60,10 @@ module RuboCop
         def autocorrect(node)
           lambda do |corrector|
             send_line = node.first_line
-            previous_line = processed_source[send_line - 2]
             next_line = processed_source[send_line]
             line = range_by_whole_lines(node.source_range)
 
-            unless previous_line_empty?(previous_line)
+            unless previous_line_empty?(send_line)
               corrector.insert_before(line, "\n")
             end
 
@@ -62,9 +81,12 @@ module RuboCop
           end
         end
 
-        def previous_line_empty?(previous_line)
+        def previous_line_empty?(send_line)
+          previous_line = previous_line_ignoring_comments(processed_source,
+                                                          send_line)
+
           block_start?(previous_line) ||
-            class_def?(previous_line) ||
+            class_def?(send_line) ||
             previous_line.blank?
         end
 
@@ -74,15 +96,16 @@ module RuboCop
 
         def empty_lines_around?(node)
           send_line = node.first_line
-          previous_line = previous_line_ignoring_comments(processed_source,
-                                                          send_line)
+
           next_line = processed_source[send_line]
 
-          previous_line_empty?(previous_line) && next_line_empty?(next_line)
+          previous_line_empty?(send_line) && next_line_empty?(next_line)
         end
 
         def class_def?(line)
-          line =~ /^\s*(class|module)/
+          return false unless @class_or_module_def_line
+
+          line == @class_or_module_def_line + 1
         end
 
         def block_start?(line)
@@ -94,10 +117,11 @@ module RuboCop
         end
 
         def message(node)
-          previous_line = processed_source[node.first_line - 2]
+          send_line = node.first_line
+          previous_line = processed_source[send_line - 2]
 
           if block_start?(previous_line) ||
-             class_def?(previous_line)
+             class_def?(send_line)
             format(MSG_AFTER, modifier: node.loc.selector.source)
           else
             format(MSG_BEFORE_AND_AFTER, modifier: node.loc.selector.source)